setup SPA routing with dummy pages
