+++ frontend/src/pages/HowToPage.js
@@ -1,0 +1,19 @@
+import React, { Component } from "react";
+import axios from "axios";
+
+class HowToPage extends Component {
+
+  render() {
+    return (
+      <main>
+        <h1>Todo app</h1>
+        <span>
+          Dummy HowTo Page
+        </span>
+      </main>
+    );
+  }
+
+}
+
+export default HowToPage;
