+++ frontend/src/pages/StatsPage.js
@@ -1,0 +1,19 @@
+import React, { Component } from "react";
+import axios from "axios";
+
+class StatsPage extends Component {
+
+  render() {
+    return (
+      <main>
+        <h1>Todo app</h1>
+        <span>
+          Dummy Stats Page
+        </span>
+      </main>
+    );
+  }
+
+}
+
+export default StatsPage;
