+++ frontend/src/pages/UploadPage.js
@@ -1,0 +1,19 @@
+import React, { Component } from "react";
+import axios from "axios";
+
+class UploadPage extends Component {
+
+  render() {
+    return (
+      <main>
+        <h1>Todo app</h1>
+        <span>
+          Dummy Upload Page
+        </span>
+      </main>
+    );
+  }
+
+}
+
+export default UploadPage;
